requirements.txt
asgiref==3.6.0
Django==4.1.6
Pillow==9.4.0
postgres==4.0
psycopg2-binary==2.9.5
psycopg2-pool==1.1
sqlparse==0.4.3
tzdata==2022.7
